+++ src/colorModels/lch.ts
@@ -1,6 +1,5 @@
 import { RgbaColor, InputObject, LchaColor } from "../types";
-import { ALPHA_PRECISION } from "../constants";
-import { clamp, clampHue, isPresent, round, roundHue } from "../helpers";
+import { clamp, clampHue, isPresent, round, roundAlpha, roundHue } from "../helpers";
 import { labaToRgba, rgbaToLaba } from "./lab";
 
 /**
@@ -19,7 +18,7 @@ export const roundLcha = (laba: LchaColor): LchaColor => ({
   l: round(laba.l, 2),
   c: round(laba.c, 2),
   h: roundHue(laba.h, 2),
-  a: round(laba.a, ALPHA_PRECISION),
+  a: roundAlpha(laba.a),
 });
 
 export const parseLcha = ({ l, c, h, a = 1 }: InputObject): RgbaColor | null => {