+++ src/constants.ts
@@ -1,10 +1,4 @@
 /**
- * We used to work with 2 digits after the decimal point, but it wasn't accurate enough,
- * so the library produced colors that were perceived differently.
- */
-export const ALPHA_PRECISION = 3;
-
-/**
  * Valid CSS <angle> units.
  * https://developer.mozilla.org/en-US/docs/Web/CSS/angle
  */